+++ src/common/global.ts
